+++ costmap_cspace/src/largemap_to_map.cpp
@@ -148,9 +148,9 @@ private:
     map.data.resize(width_ * width_);
 
     const int gx =
-        lroundf((map.info.origin.position.x - large_map_->info.origin.position.x) / map.info.resolution);
+        std::lround((map.info.origin.position.x - large_map_->info.origin.position.x) / map.info.resolution);
     const int gy =
-        lroundf((map.info.origin.position.y - large_map_->info.origin.position.y) / map.info.resolution);
+        std::lround((map.info.origin.position.y - large_map_->info.origin.position.y) / map.info.resolution);
     const float half_width = width_ / 2.0;
 
     for (int y = gy; y < gy + width_; ++y)
